+++ index.html
@@ -37,6 +37,10 @@ body { margin: 0; overflow: hidden; background:#000; touch-action: none;}
 <div id="message"></div>
 <script src="https://cdn.jsdelivr.net/npm/three@0.157.0/build/three.min.js"></script>
 <script src="cityLights.js"></script>
+<script type="module" src="car.js"></script>
+<script type="module" src="track.js"></script>
+<script type="module" src="ai.js"></script>
+<script type="module" src="powerups.js"></script>
 <script type="module" src="game.js"></script>
 </body>
 </html>
